+++ aioresult/__init__.py
@@ -1,7 +1,11 @@
+# Copyright Arthur Tacca 2022 - 2024
+# Distributed under the Boost Software License, Version 1.0.
+# See accompanying file LICENSE or the copy at https://www.boost.org/LICENSE_1_0.txt
 
 from aioresult._version import __version__
 
 from aioresult._src import (
-    ResultBase, ResultCapture, StartableResultCapture, Future,
+    ResultBase, ResultCapture, Future,
     TaskFailedException, TaskNotDoneException, FutureSetAgainException
 )
+from aioresult._wait import wait_all, wait_any, results_to_channel