+++ docs/source/conf.py
@@ -124,19 +124,14 @@ html_theme_path = [sphinx_rtd_theme.get_html_theme_path()]
 # documentation.
 #
 html_theme_options = {
-    # default is 2
-    # show deeper nesting in the RTD theme's sidebar TOC
-    # https://stackoverflow.com/questions/27669376/
-    # I'm not 100% sure this actually does anything with our current
-    # versions/settings...
-    "navigation_depth": 4,
-    "logo_only": True,
+    "navigation_depth": 2,
+    "collapse_navigation": False
 }
 
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
-html_static_path = ['_static']
+html_static_path = []
 
 
 # -- Options for HTMLHelp output ------------------------------------------